+++ test/net/talvi/puffinplot/data/DatumTest.java
@@ -451,10 +451,12 @@ public class DatumTest {
     }
     
     @Test
-    public void testToSummaryString() {
+    public void testToSummaryStringThermalAndDegauss() {
         final Sample sample = new Sample("test", null);
-        final Datum d0 = makeSimpleDatum(TreatType.THERMAL, 75, 24.1, -23.2, 7.123e-3);
-        final Datum d1 = makeSimpleDatum(TreatType.DEGAUSS_XYZ, 0.35, 0.0, 3.5, 1.45e-2);
+        final Datum d0 = makeSimpleDatum(TreatType.THERMAL,
+                75, 24.1, -23.2, 7.123e-3);
+        final Datum d1 = makeSimpleDatum(TreatType.DEGAUSS_XYZ,
+                0.35, 0.0, 3.5, 1.45e-2);
         sample.addDatum(d0);
         sample.addDatum(d1);
         d0.setMagSus(1.234);
@@ -464,7 +466,24 @@ public class DatumTest {
         assertEquals("Step 2/2 | 3-axis degauss, 350 mT | "
                 + "Dec: 0.0° | Inc: 3.5° | Mag: 1.45e−02 A/m",
                 d1.toSummaryString());
-        
+    }
+    
+    @Test
+    public void testToSummaryStringNoneAndIrm() {
+        final Sample sample = new Sample("test", null);
+        final Datum d0 = makeSimpleDatum(TreatType.NONE,
+                75, 24.1, -23.2, 7.123e-3);
+        final Datum d1 = makeSimpleDatum(TreatType.IRM,
+                0.35, 0.0, 3.5, 1.45e-2);
+        sample.addDatum(d0);
+        sample.addDatum(d1);
+        d0.setMagSus(1.234);
+        assertEquals("Step 1/2 | No treatment | "
+                + "Dec: 24.1° | Inc: −23.2° | Mag: 7.12e−03 A/m | MS: 1.23",
+                d0.toSummaryString());
+        assertEquals("Step 2/2 | IRM, 350 mT | "
+                + "Dec: 0.0° | Inc: 3.5° | Mag: 1.45e−02 A/m",
+                d1.toSummaryString());
     }
     
     private static Datum makeSimpleDatum(TreatType treatType,
@@ -472,16 +491,24 @@ public class DatumTest {
         final Datum d = new Datum();
         d.setTreatType(treatType);
         switch (treatType) {
+            case NONE:
+            case UNKNOWN:
+                break;
             case THERMAL:
                 d.setTemp(level);
                 break;
             case IRM:
                 d.setIrmField(level);
                 break;
-            default:
+            case DEGAUSS_XYZ:
                 d.setAfX(level);
                 d.setAfY(level);
+                // fall-through intentional
+            case DEGAUSS_Z:
                 d.setAfZ(level);
+                break;
+            default:
+                throw new Error("Unhandled treatment type "+treatType);
                 
         }
         d.setMoment(Vec3.fromPolarDegrees(moment, inc, dec));